+++ cloudflare/structs/resp/updateLicense.go
@@ -7,10 +7,10 @@ type UpdateLicenseData struct {
 	AccountType              string    `json:"account_type"`
 	Created                  time.Time `json:"created"`
 	Updated                  time.Time `json:"updated"`
-	PremiumData              int       `json:"premium_data"`
-	Quota                    int       `json:"quota"`
+	PremiumData              int64     `json:"premium_data"`
+	Quota                    int64     `json:"quota"`
 	WarpPlus                 bool      `json:"warp_plus"`
-	ReferralCount            int       `json:"referral_count"`
-	ReferralRenewalCountdown int       `json:"referral_renewal_countdown"`
+	ReferralCount            int64     `json:"referral_count"`
+	ReferralRenewalCountdown int64     `json:"referral_renewal_countdown"`
 	Role                     string    `json:"role"`
 }
